Works in cloudfoundry

+++ src/main/java/ie/cit/cloudapp/JdbcPlayerRepository.java
@@ -8,7 +8,10 @@ import java.util.List;
 
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.RowMapper;
+import org.springframework.security.access.annotation.Secured;
+import org.springframework.security.core.context.SecurityContextHolder;
 
+@Secured("ROLE_USER")
 public class JdbcPlayerRepository {
 	
 	private JdbcTemplate jdbcTemplate;
@@ -22,23 +25,28 @@ public class JdbcPlayerRepository {
 	}
 	
 	public void save(Player player) {
-		jdbcTemplate.update("insert into store(name, gk, rb, cb1, cb2, lb, rw, cm1, cm2, lw, st1, st2, done) values(?,?,?,?,?,?,?,?,?,?,?,?,?)",
-				player.getName(), player.getGk(), player.getRb(), player.getCb1(), player.getCb2(), player.getLb(),player.getRw(), player.getCm1(), player.getCm2(), player.getLw(), player.getSt1(), player.getSt2(), player.isDone());
+		jdbcTemplate.update("insert into store(name, gk, rb, cb1, cb2, lb, rw, cm1, cm2, lw, st1, st2, done, owner) values(?,?,?,?,?,?,?,?,?,?,?,?,?,?)",
+				player.getName(), player.getGk(), player.getRb(), player.getCb1(), player.getCb2(), player.getLb(),player.getRw(), player.getCm1(), player.getCm2(), player.getLw(), player.getSt1(), player.getSt2(), player.isDone(), getCurrentUser());
 	}
 	
 	public Player get(int id) {
 		return jdbcTemplate.queryForObject(
-				"select id, name, gk, rb, cb1, cb2, lb, rw, cm1, cm2, lw, st1, st2, done from store where id=?", new PlayerMapper(),
-				id);
+				"select id, name, gk, rb, cb1, cb2, lb, rw, cm1, cm2, lw, st1, st2, done from store where id=? and owner=?", new PlayerMapper(),
+				id, getCurrentUser());
 	}
 		
 		public List<Player> getAll() {
-			return jdbcTemplate.query("select id, name, gk, rb, cb1, cb2, lb, rw, cm1, cm2, lw, st1, st2, done from store",
-					new PlayerMapper());
+			return jdbcTemplate.query("select id, name, gk, rb, cb1, cb2, lb, rw, cm1, cm2, lw, st1, st2, done from store where owner=?",
+					new PlayerMapper(), getCurrentUser());
+		}
+
+		private String getCurrentUser() {
+			return SecurityContextHolder.getContext().getAuthentication().getName();
 		}
 		
 		public void delete(int id) {
-			jdbcTemplate.update("delete from store where id=?", id);
+			jdbcTemplate.update("delete from store where id=? and owner=?", id,
+					getCurrentUser());
 	}
 		
 		class PlayerMapper implements RowMapper<Player> {
